+++ lib/config/rules/shopify.js
@@ -37,6 +37,8 @@ module.exports = {
   'shopify/prefer-twine': 'error',
   // Require that React component state be initialized when it has a non-empty type.
   'shopify/react-initialize-state': 'off',
+  // Disallow multiple render methods in React component classes.
+  'shopify/react-no-multiple-render-methods': 'off',
   // Prefer all non-React-specific members be marked private in React class components.
   'shopify/react-prefer-private-members': 'off',
   // Require that React component state be typed in TypeScript.
