+++ lib/rules/react-no-multiple-render-methods.js
@@ -1,13 +1,21 @@
 const Components = require('eslint-plugin-react/lib/util/Components');
 
+const message = [
+  'Don’t use multiple render methods in a single component;',
+  'they generally make your component harder to read.',
+  'Instead break {{name}} out into its own component',
+  'and render it inside this one.',
+].join(' ');
+
 module.exports = {
   meta: {
     docs: {
-      description: 'Disallow renderX methods within React component classes',
+      description:
+        'Disallow multiple render methods in React component classes',
       category: 'Best Practices',
       recommended: true,
       uri:
-        'https://github.com/Shopify/eslint-plugin-shopify/blob/master/docs/rules/react-no-render-methods.md',
+        'https://github.com/Shopify/eslint-plugin-shopify/blob/master/docs/rules/react-no-multiple-render-methods.md',
     },
   },
 
@@ -21,7 +29,7 @@ module.exports = {
 
       context.report({
         node,
-        message: `No renderX methods. '{{name}}' should be a seperate component.`,
+        message,
         data: {name},
       });
     }
@@ -30,6 +38,7 @@ module.exports = {
       ClassDeclaration(node) {
         isES6Component = utils.isES6Component(node);
       },
+
       MethodDefinition(node) {
         if (!isES6Component || isRenderMethod(node)) {
           return;
